Added clustering model script.
This script clusters (n_clusters=8) the training set in the X,Y dimensions. For each cluster, a count of each category is done to provide a vector of relative probabilities for each category. For the predictions, each test item's (X,Y) location is fed into the clustering model to determine which cluster it belongs to. The test item recieves the probs for the predicted cluster.