Works in principle, but need to deal better with large training set.
This kNN code works in principal but hangs on large training sets (using sample of the 800k+ provided training obsvervations). Unfortunately, without a large training set, there are not enough instances of rare crime types like TREA (treason?) and we get mismatches in target dimension in the cross-validation calculations.

Possibilities for improvement:

1.) Maybe do CV by hand instead of using the scikit functions?

2.) Take out rare crimes and assign low probs by hand (maybe even zero prob). Train a kNN classifier on the remaining non-rare-crime data set. Use this classifier to predict probs (for the non-rare crimes) on the test data. Sew these probs together with the hand-set ones for the rare crimes (Need to normalize all the non-rare probs by multiplying by 1-p, where p is total hand-set probs of the rare crimes).